Pull request #137 introduced a regression as it added a significant delay in job processing. (#144)
Previously NewItemInQueueEvent was raised within the current pending transaction. Therefore the queue could not retrieve the added job, as it was not visible to the queue yet.

Changed the logic so that NewItemInQueueEvent gets raised **after** transaction scope is completed.

TransactionCompleted event is required here, because if this TransactionScope is enlisted within an ambient TransactionScope, the ambient TransactionScope controls when the TransactionScope completes.